projects/05_capstone_project/requirements.txt
aniso8601==6.0.0
alembic==1.5.4
Babel==2.9.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3
## The following requirements were added by pip freeze:
astroid==2.2.5
atomicwrites==1.4.0
attrs==20.3.0
awscli==1.19.49
backcall==0.2.0
botocore==1.20.49
colorama==0.4.3
decorator==4.4.2
docutils==0.15.2
ecdsa==0.13.2
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
future==0.17.1
gunicorn==20.0.4
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==4.3.18
jedi==0.18.0
jmespath==0.10.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.4.0
mccabe==0.6.1
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.17
py==1.10.0
pyasn1==0.4.8
pycryptodome==3.3.1
Pygments==2.8.1
PyJWT==1.4.2
pylint==2.3.1
pyparsing==2.4.7
pytest==6.2.2
python-jose==3.2.0
python-jose-cryptodome==1.3.2
#pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
rsa==4.7.2
s3transfer==0.3.6
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
urllib3==1.26.4
wcwidth==0.2.5
wrapt==1.11.1
Flask-Script==2.0.6